testproject/pyproject.toml
[tool.poetry]
name = "testproject"
version = "0.1.0"
description = ""
authors = ["Jonhnatha Trigueiro <joepreludian@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
pytest-django = "^4.0.0"
pytest-cov = "^2.10.1"
prelude-django-admin-toolkit = {path = "../", develop=true}
ipdb = "^0.13.4"
pydevd = "^2.2.0"

[tool.poetry.dev-dependencies]
neovim = "^0.3.1"
behave-django = "^1.4.0"
splinter = "^0.13.0"
django-livereload-server = "^0.3.2"
honcho = "^1.0.1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
